pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markitdown-gui"
dynamic = ["version"]
description = "Windows desktop GUI for converting PDF, DOCX, XLSX, and HTML files to Markdown"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["markitdown", "gui", "pyside6", "windows", "markdown"]
authors = [
  { name = "kamishioya" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "markitdown[pdf,docx,xlsx]>=0.1.6b2,<0.2.0",
  "PySide6_Essentials>=6.8.0,<7.0.0",
]

[project.optional-dependencies]
build = [
  "pyinstaller>=6.11.0,<7.0.0",
]

[project.urls]
Documentation = "https://github.com/kamishioya/markitdown-gui#readme"
Issues = "https://github.com/kamishioya/markitdown-gui/issues"
Source = "https://github.com/kamishioya/markitdown-gui"
Upstream = "https://github.com/microsoft/markitdown"

[tool.hatch.version]
path = "src/markitdown_gui/__about__.py"

[project.scripts]
markitdown-gui = "markitdown_gui.__main__:main"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
  "PySide6_Essentials>=6.8.0,<7.0.0",
]

[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/markitdown_gui tests}"

[tool.coverage.run]
source_pkgs = ["markitdown-gui", "tests"]
branch = true
parallel = true
omit = [
  "src/markitdown_gui/__about__.py",
]

[tool.coverage.paths]
markitdown-gui = ["src/markitdown_gui", "*/markitdown-gui/src/markitdown_gui"]
tests = ["tests", "*/markitdown-gui/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.build.targets.sdist]
only-include = ["src/markitdown_gui"]
